Corrected Lock mem cache commit call
The commit call is writing a 0 when the cache has a 1 for StreamVersion.
When the tape is read to load up the cache, the Stream versions are all
off by 1. You end up with a concurrency exception when you fire up the
system and try to handle a command on an aggregate that already exists.